fix: add text-align and update placeholder
