+++ includes/blocks/featured-image-caption/block.json
@@ -25,7 +25,8 @@
 		},
 		"typography": {
 			"fontSize": true,
-			"lineHeight": true
+			"lineHeight": true,
+			"textAlign": true
 		}
 	}
 }
